commonline-core/src/main/java/commonline/core/util/Base36Converter.java
/**
 * Copyright 2008-2009 the original author or authors.
 * 
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in
 * compliance with the License. You may obtain a copy of the License at:
 * 
 * http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software distributed under the License is
 * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and limitations under the License.
 */
package commonline.core.util;

public class Base36Converter {

    public String convertTo(int base10Value) {
        return Integer.toString(base10Value, 36).toUpperCase();
    }

    public int convertFrom(String base36Value) {
        return Integer.parseInt(base36Value, 36);
    }
}
